refactored some of the js
