+++ index.html
@@ -158,7 +158,7 @@
 <div class="custom-cursor site-wide">
     <img src="/Resources/cursor-default.png" id="cursorDefault"></img>
 </div>
-<script src="/js/cursor.js"></script>
+<script src="/js/script.js" type="module"></script>
 <script src="/js/slide.js"></script>
 </body>
 </html>
