+++ wwwroot/css/poke-styles.css
@@ -9,6 +9,39 @@ body {
   font-family: Arial, Helvetica, sans-serif;
 }
 
+/* Removes default cursor*/
+* {
+  cursor: none !important;
+}
+
+/* Positions the cursor*/
+.custom-cursor {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 30px;
+  height: 30px;
+  display: none;
+  pointer-events: none;
+}
+
+.site-wide img{
+  max-width: 100%;
+  max-height: 100%;
+}
+
+.site-wide.active {
+  width: 35px;
+  height: 35px;
+}
+
+body:has(a:hover) #cursorDefault {
+  filter: drop-shadow(2px 0 0 #ffffff) 
+      drop-shadow(0 2px 0 #ffffff)
+      drop-shadow(-2px 0 0 #ffffff) 
+      drop-shadow(0 -2px 0 #ffffff);
+}
+
 header {
   border-bottom: 2px solid black;
   padding: 10px;
