+++ wwwroot/js/pokemon-game/gameModes.js
