+++ wwwroot/js/pokemon-game/main.js
@@ -1,11 +1,13 @@
 import { setLogoAndPochama } from "./style/pochama.js";
-import { startGameLoop } from "./game/game.js";
+import { startGameLoop } from "./game.js";
+import { addCursor } from "../cursor.js"; 
 
 // gameStart button
 const ranButton = document.getElementsByClassName("random-button")[0];
 
 // run at startup Methods
 setLogoAndPochama();
+addCursor();
 
 // event handlers
 ranButton.addEventListener("click", async () => {
