+++ wwwroot/js/pokemon-game/objectAndFetch/fetchTokenAndScore.js
