+++ wwwroot/js/pokemon-game/objectAndFetch/pokefactory.js
