+++ wwwroot/js/pokemon-game/questions.js
