+++ wwwroot/js/pokemon-game/submitScore.js
@@ -1,7 +1,6 @@
 import {
-  getTokenOnGameStart,
   submitScore,
-} from "../objectAndFetch/fetchTokenAndScore.js";
+} from "./objectAndFetch/fetchTokenAndScore.js";
 
 const submit = async (playerName, score, token) => {
   await submitScore(playerName, score, token);
