+++ wwwroot/js/script.js
@@ -1,0 +1,4 @@
+import { addCursor } from "./cursor.js";
+
+// run at startup
+addCursor();
