+++ wwwroot/leaderboard.html
@@ -9,7 +9,7 @@
     <link rel="icon" href="" id="logo">
     <meta name="author" content="HelloHaruSystem">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="/js/leaderboard.js" type="module" defer></script>
+    <script src="/js/pokemon-game/leaderboard.js" type="module" defer></script>
 </head>
 
 <body>
@@ -63,5 +63,9 @@
             </ul>
     </div>
 </footer>
+<!--Used for cursor-->
+<div class="custom-cursor site-wide">
+    <img src="/Resources/cursor-default.png" id="cursorDefault"></img>
+</div>
 </body>
 </html>
